Some improvements
Added z, facing, and pitch labels
Moved FPS and version labels to misc
Use the correct type for setting config values (fixes the config gui not loading after using /spk df)
No more completely unnecessary usage of ArrayList in LabelManager
Made the help command message use config colors instead of being hardcoded

+++ src/main/java/me/oondanomala/spkmod/config/Config.java
@@ -40,7 +40,7 @@ public class Config {
         configuration.save();
     }
 
-    public void setConfigOption(String optionName, double value) {
+    public void setConfigOption(String optionName, int value) {
         SPKMod.config.configuration.getCategory(Configuration.CATEGORY_CLIENT).get(optionName).set(value);
         if (configuration.hasChanged()) {
             configuration.save();
@@ -53,7 +53,6 @@ public class Config {
             configuration.save();
         }
     }
-
 
     @SubscribeEvent
     public void onConfigChange(ConfigChangedEvent.OnConfigChangedEvent event) {
